move mock listener definitions to where its used
